Add responsiveness to the datatables
