+++ packages/webc/package.json
@@ -13,7 +13,8 @@
   },
   "dependencies": {
     "bootstrap": "^5.1.3",
-    "datatables.net-bs5": "^1.11.3",
+    "datatables.net-bs5": "^1.13.4",
+    "datatables.net-responsive-bs5": "^2.4.1",
     "firebase": "^9.4.1",
     "firebase-admin": "^9.5.0",
     "jquery": "^3.6.0",
